+++ src/ssl/Server.cpp
@@ -187,7 +187,7 @@ Server::send(const ID id, const uint32_t len, const uint8_t* const data,
    * Write the data. With BIO mem, the write will always succeed.
    */
   off = 0;
-  off += SSL_write(c.ssl, data, len);
+  off += SSL_write(c.ssl, data, (int)len);
   /*
    * Flush the data.
    */
